read vmd file

+++ Chapter 10/Chapter 10/BasicVertexShader.hlsl
@@ -15,8 +15,8 @@ Output main (float4 pos : POSITION, float4 normal : NORMAL, float2 uv : TEXCOORD
 	Output output;
 	
 	// about bone.
-	float w = weight / 100.0f;
-	matrix bm = bones[boneno[0]] * w + bones[boneno[1]] * (1 - w);
+	float w = (float)weight / 100.0f;
+	matrix bm = bones[boneno[0]] * w + bones[boneno[1]] * (1.0 - w);
 	pos = mul (bm, pos);
 	
 	pos = mul (world, pos);
